tascam-runtime: code refactoring for runtime structure
